Update docs and demos
