+++ demos/dqx_ldp_demo_foreach_batch_quarantine.py
@@ -5,26 +5,26 @@ from pyspark import pipelines as dp
 
 # MAGIC
 # MAGIC %md
-# MAGIC ## DQX in a Lakeflow Pipeline using a foreachBatch sink — data quality with quarantine pattern
+# MAGIC ## Using DQX in a Lakeflow Declarative Pipeline with a foreachBatch sink — data quality with quarantine pattern
 # MAGIC
-# MAGIC This demo applies DQX checks in a streaming Lakeflow pipeline and **splits** each micro-batch into a
-# MAGIC valid `silver` table and a `quarantine` table. It also appends **per-batch summary metrics** — one set
-# MAGIC of counts computed and appended for each micro-batch to a `dq_summary_metrics` table.
+# MAGIC This demo applies DQX checks and **splits** each micro-batch into a valid `silver` table and a
+# MAGIC `quarantine` table. Summary metrics are then computed for each incremental micro-batch and persisted to a
+# MAGIC table using a [foreachBatch sink](https://docs.databricks.com/aws/en/ldp/for-each-batch).
+# MAGIC The metrics table contains **per-batch summary metrics** about the input data quality.
 # MAGIC
-# MAGIC **When to use this (foreachBatch sink) instead of the simple quarantine pipeline (`dqx_dlt_demo_quarantine.py`):**
-# MAGIC use it when you want summary metrics computed **incrementally per micro-batch** and appended as a
-# MAGIC history (one row set per batch), rather than the cumulative full-table snapshot the materialized-view
-# MAGIC demo produces. This is also **potentially more performant on large or growing tables**: metrics are
-# MAGIC aggregated over only the current micro-batch, whereas the materialized view re-aggregates over the
-# MAGIC whole checked table on each pipeline update. If you just need to split records into valid and
-# MAGIC quarantine tables, prefer the simpler `dqx_dlt_demo_quarantine.py`.
+# MAGIC **When to use this a foreachBatch sink:**
+# MAGIC Use a foreachBatch sink when you want summary metrics computed **incrementally per micro-batch** and
+# MAGIC appended as a history with one set of rows per batch, rather than a cumulative full-table snapshot. This
+# MAGIC may be **more performant for large or growing tables**. While a materialized view re-aggregates over the
+# MAGIC whole checked table on each pipeline update, this example aggregates metrics over only the current micro batch.
+# MAGIC If you just need to split records into valid and quarantine tables, prefer the simpler `dqx_ldp_demo_quarantine.py`.
 # MAGIC
-# MAGIC For the non-quarantine variant of this pattern, see `dqx_dlt_demo_foreach_batch.py`.
+# MAGIC To see how to use this pattern without quarantining data, see `dqx_ldp_demo_foreach_batch.py`.
 # MAGIC
-# MAGIC Create new ETL Pipeline to execute this notebook (see [here](https://docs.databricks.com/aws/en/getting-started/data-pipeline-get-started)):
+# MAGIC Create a new ETL Pipeline to execute this notebook (see [here](https://docs.databricks.com/aws/en/getting-started/data-pipeline-get-started)):
 # MAGIC 1. Upload the notebook to a Databricks Workspace
 # MAGIC 2. Go to `Workflows` tab > `Create` > `ETL Pipeline` > `Add existing assets` > select the source code path and root directory
-# MAGIC 3. Add DQX library as a [dependency](https://docs.databricks.com/aws/en/dlt/dlt-multi-file-editor#environment) to the pipeline: Go to `Settings` > `Edit environment` > Add `databricks‑labs‑dqx` as dependency
+# MAGIC 3. Add the DQX library as a [dependency](https://docs.databricks.com/aws/en/dlt/dlt-multi-file-editor#environment) to the pipeline: Go to `Settings` > `Edit environment` > Add `databricks‑labs‑dqx` as a dependency
 # MAGIC 4. Run the pipeline
 # MAGIC
 # MAGIC
@@ -42,23 +42,21 @@ from databricks.labs.dqx.engine import DQEngine
 from databricks.labs.dqx.metrics_observer import DQMetricsObserver
 from databricks.sdk import WorkspaceClient
 
-# compute_summary_metrics requires an observer on the engine; it reads any custom_metrics from it.
+# Create the DQEngine with a metrics observer to get summary metrics
 dq_engine = DQEngine(WorkspaceClient(), observer=DQMetricsObserver())
 
-# A foreachBatch sink writes to tables *outside* the pipeline, so it must use fully-qualified names
-# (unqualified names would be captured as pipeline-managed streaming tables and rejected). The target
-# catalog and schema are read from the pipeline configuration, defaulting to the demo schema.
+# Read the target catalog and schema are read from the pipeline configuration
 output_catalog = spark.conf.get("demo_catalog", "main")
-output_schema = spark.conf.get("demo_schema", "dqx_dlt_demo_foreach_batch_quarantine")
+output_schema = spark.conf.get("demo_schema", "dqx_ldp_demo_foreach_batch_quarantine")
 
 # COMMAND ----------
 
 # MAGIC %md
 # MAGIC ## Define Data Quality checks
+# MAGIC Define your quality checks in YAML. These can also be defined using DQX classes or loaded from a table of file.
 
 # COMMAND ----------
 
-# Define checks in YAML format. They can also be defined using classes or loaded from a file or a table.
 checks = yaml.safe_load("""
 - check:
     function: is_not_null
@@ -124,7 +122,8 @@ checks = yaml.safe_load("""
 # COMMAND ----------
 
 # MAGIC %md
-# MAGIC ## Define Lakeflow Pipeline (bronze -> foreachBatch sink -> silver + quarantine + metrics)
+# MAGIC ## Bronze Layer
+# MAGIC Stream the source data from a Delta table and write it to a bronze-layer table.
 
 # COMMAND ----------
 
@@ -136,27 +135,41 @@ def bronze():
 
 # COMMAND ----------
 
-# foreachBatch sink: apply checks per micro-batch, split into valid and quarantine rows, and append
-# summary metrics for that batch. Computing metrics inside the sink gives a per-batch history, unlike a
-# @dp.table materialized view which recomputes a cumulative snapshot over the whole table.
-# The sink writes to fully-qualified tables outside the pipeline (see output_catalog/output_schema above).
+# MAGIC %md
+# MAGIC ## foreachBatch Sink
+# MAGIC Define a foreachBatch sink that applies quality checks using `dq_engine.apply_checks_by_metadata_and_split`,
+# MAGIC splits each micro-batch into valid and quarantine rows and appends them to `silver` and `quarantine` tables,
+# MAGIC computes the quality metrics for the incremental batch, and appends the computed metrics to a metrics table.
+# MAGIC
+# MAGIC Computing metrics with a foreachBatch sink creates a per-batch history (unlike a materialized view
+# MAGIC defined using `@dp.table`, which recomputes a snapshot of the metrics over the entire table). This is useful
+# MAGIC for tracking data quality of different batches and avoiding dataset-level aggregation over very large tables.
+
+# COMMAND ----------
+
 @dp.foreach_batch_sink(name="silver_quarantine_sink")
 def silver_quarantine_sink(batch_df, batch_id):
-  # Split into valid rows (auxiliary result columns dropped) and invalid rows (errors or warnings).
+  # Split into valid rows (auxiliary result columns dropped) and invalid rows with errors or warnings, and append each to its table
   valid_df, quarantine_df = dq_engine.apply_checks_by_metadata_and_split(batch_df, checks)
   valid_df.write.format("delta").mode("append").saveAsTable(f"{output_catalog}.{output_schema}.silver")
   quarantine_df.write.format("delta").mode("append").saveAsTable(f"{output_catalog}.{output_schema}.quarantine")
 
-  # Summary metrics: computed over the full checked batch (before the split) and appended per batch.
-  # Note: with a foreachBatch sink these are per-batch counts (one set of rows per micro-batch), not a
-  # cumulative snapshot; aggregate across batches downstream if you need running totals.
+  # Compute the summary metrics over the full checked batch (before the split) and append them to the metrics table
+  # Note: A foreachBatch sink computes metrics as per-batch counts (one set of rows per micro-batch), not a
+  # cumulative snapshot; Aggregate across batches downstream if you need running totals
   checked_df = dq_engine.apply_checks_by_metadata(batch_df, checks)
   metrics_df = dq_engine.compute_summary_metrics(checked_df, checks=checks)
   metrics_df.write.format("delta").mode("append").saveAsTable(f"{output_catalog}.{output_schema}.dq_summary_metrics")
 
+
 # COMMAND ----------
 
-# Wire the streaming bronze table into the sink via an append flow.
+# MAGIC %md
+# MAGIC ## Silver, Quarantine, and Metrics Layer
+# MAGIC Use `@dp.append_flow` to read the bronze data and run the foreachBatch sink over each micro-batch.
+
+# COMMAND ----------
+
 @dp.append_flow(target="silver_quarantine_sink")
 def silver_quarantine_flow():
   return spark.readStream.table("bronze")
