+++ docs/dqx/docs/demos.mdx
@@ -18,10 +18,10 @@ Import the following notebooks in the Databricks workspace to try DQX out:
 * [DQX Demo Notebook for Data Contract Integration (ODCS)](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_demo_datacontract_odcs.py) - demonstrates how to generate DQX quality rules from ODCS (Open Data Contract Standard) data contracts, including predefined rules from schema constraints, explicit custom rules, and contract metadata tracking.
 * [DQX Demo Notebook for Spark Structured Streaming (Native End-to-End Approach)](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_streaming_demo_native.py) - demonstrates how to use DQX as a library with Spark Structured Streaming, using the built-in end-to-end method to handle both reading and writing.
 * [DQX Demo Notebook for Spark Structured Streaming (DIY Approach)](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_streaming_demo_diy.py) - demonstrates how to use DQX as a library with Spark Structured Streaming, while handling reading and writing on your own outside DQX using Spark API.
-* [DQX Demo Notebook for Lakeflow Pipelines (formerly DLT)](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_dlt_demo.py) - demonstrates how to use DQX in a Lakeflow Pipeline, reporting issues as additional columns and computing data quality summary metrics.
-* [DQX Demo Notebook for Lakeflow Pipelines with Quarantine](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_dlt_demo_quarantine.py) - demonstrates how to use DQX in a Lakeflow Pipeline to split valid and invalid records into valid and quarantine tables and compute summary metrics.
-* [DQX Demo Notebook for Lakeflow Pipelines with a foreachBatch sink](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_dlt_demo_foreach_batch.py) - demonstrates how to use DQX in a Lakeflow Pipeline foreachBatch sink, reporting issues as additional columns and computing summary metrics incrementally.
-* [DQX Demo Notebook for Lakeflow Pipelines with a foreachBatch sink](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_dlt_demo_foreach_batch_quarantine.py) - demonstrates how to use DQX in a Lakeflow Pipeline foreachBatch sink to split valid and invalid records into separate tables in the same batch and compute summary metrics incrementally.
+* [DQX Demo Notebook for Lakeflow Pipelines (formerly DLT)](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_ldp_demo.py) - demonstrates how to use DQX in a Lakeflow Pipeline, reporting issues as additional columns and computing data quality summary metrics.
+* [DQX Demo Notebook for Lakeflow Pipelines with Quarantine](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_ldp_demo_quarantine.py) - demonstrates how to use DQX in a Lakeflow Pipeline to split valid and invalid records into valid and quarantine tables and compute summary metrics.
+* [DQX Demo Notebook for Lakeflow Pipelines with a foreachBatch sink](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_ldp_demo_foreach_batch.py) - demonstrates how to use DQX in a Lakeflow Pipeline foreachBatch sink, reporting issues as additional columns and computing summary metrics incrementally.
+* [DQX Demo Notebook for Lakeflow Pipelines with a foreachBatch sink](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_ldp_demo_foreach_batch_quarantine.py) - demonstrates how to use DQX in a Lakeflow Pipeline foreachBatch sink to split valid and invalid records into separate tables in the same batch and compute summary metrics incrementally.
 * [DQX usage with Declarative Automation Bundles](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_demo_asset_bundle/README.md) - demonstrates how to use DQX as a library with Declarative Automation Bundles (formerly Databricks Asset Bundles).
 * [DQX usage with dbt](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_demo_dbt/README.md) - demonstrates how to use DQX as a library with dbt projects.
 * [DQX Demo on Primary-Key detection](https://github.com/databrickslabs/dqx/blob/v0.15.0/demos/dqx_demo_llm_pk_detection.py) - demonstrates how to detect primary keys and generate uniqueness rules.
